* Add build configuration CONFIG_HAVE_POSIX_SPINLOCK, as POSIX spin locks are not available on Cygwin or OS X. * Add winmm.lib as dependency on Windows in order to set higher resolution timers. * Set higher resolution timers (1ms) with timeBeginPeriod on PGM engine init, and close out on shutdown with timeEndPeriod. * Check for Cygwin when setting multicast loop, calling getnetbyname, setting pktinfo contents. * Add casts to satisfy Win64 MSVC builds. * Replace test address family from AF_IPX to AF_SNA as it is available in Cygwin. * Add pgm/in.h for OS X which lacks RFC 3768 structures for ASM & SSM joins. * Add feature macro fallbacks for Darwin. * Add basic spin lock using GCC intrinsics. * Fix Windows SRW lock implementation. * OS X does not support read-mostly segment. * Update pgm_poll_info to work with Windows WSAPoll implementation (Vista+). * Add Cygwin build environment, note that SConscripts fail on all Command statements and need manual intervention. * Replace WSAEVENT quit notification with sockets in order to use select or WSAPoll. * Add WSAPoll support. * Fix setting router alert on Win64. * Add test for FreeBSD which does not support large socket buffers. * Avoid Sleep(0) yielding on Windows XP/2000 in pgmping. * Add support for select() in pgmping. * Update pgm_poll_info to use short for event type. * Add OS X 10.6 build environment, note C++ examples disabled due to name space issues with RFC 3768 structures. * Only set TXW_MAX_RTE if values above 0 to allow disabling on command line in pgmping. * Bump version to 5.1.

+++ trunk/openpgm/pgm/packet_test.c
@@ -140,7 +140,7 @@ pgm_print_packet (
 
 /* packets that fail checksum will generally not be passed upstream except with rfc3828
  */
-	const uint16_t ip_sum = pgm_inet_checksum(data, packet_length, 0);
+	const uint16_t ip_sum = pgm_inet_checksum(data, (uint16_t)packet_length, 0);
 	if (ip_sum != 0) {
 		const uint16_t encoded_ip_sum = ntohs(ip->ip_sum);
 		printf (", bad cksum! %i", encoded_ip_sum);
